Hook keyword!
Use-package has a :hook keyword that simplifies a lot of the hooks
I've been writing. Sooo, this is a refactor of those cases.
Also, I added two more packages to my config: hl-todo (to highlight
TODO, FIXME, and so on) and js2-refactor that I thought I'd included
in the previuos commit